WIP on admin filtering so I don't lose changes while I check out logs branch for reference
